+++ mRemoteNGTests/Config/Serializers/ConnectionSerializers/Csv/CsvConnectionsDeserializerMremotengFormatTests.cs
@@ -99,7 +99,7 @@ namespace mRemoteNGTests.Config.Serializers.ConnectionSerializers.Csv
                 UseCredSsp = true,
                 UseVmId = false,
                 UseEnhancedMode = false,
-                RenderingEngine = HTTPBase.RenderingEngine.Gecko,
+                RenderingEngine = HTTPBase.RenderingEngine.CEF,
                 ICAEncryptionStrength = IcaProtocol.EncryptionStrength.Encr40Bit,
                 RDPAuthenticationLevel = AuthenticationLevel.WarnOnFailedAuth,
                 Colors = RDPColors.Colors16Bit,