+++ CHANGELOG.md
@@ -12,7 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Changed
 - #1767: Turned about window into a simple popup form
 - #1766: Converted components check page into options page
-- #1460: Updated GeckoFX to v60
+- #1690: Replaced GeckoFX (Firefox) with CefSharp (Chromium)
 ### Fixed
 - #1773: Connection issue with mysql - Missing fields in tblCons
 - #1713: Sound redirection does not work if Clipboard redirection is set to No